# Errors() +performance : using strings.Builder instead of string concatenation