@manisharma 2023-09-29
> Thanks, makes perfect sense to change that. The `strings.Builder` didn't exist back when I was writing this code.
> 
> No need to include 2 versions and benchmarks, though, we're not a strings building library ;-). Just change the old implementation to the new one.

Done, thanks!